% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  alleles,
  do_filter,
  fasta_filename,
  folder_name,
  mhc_haplotype,
  netmhc2pan_archive_filename,
  netmhc2pan_bin_tarfile_path,
  netmhc2pan_bin_url,
  netmhc2pan_data_tarfile_path,
  netmhc2pan_data_url,
  netmhc2pan_folder_name,
  netmhc2pan_subfolder,
  netmhc2pan_version,
  os,
  peptide_length,
  peptides,
  protein_sequence,
  temp_fasta_filename,
  temp_local_file,
  temp_xls_filename,
  verbose,
  xls_filename
)
}
\arguments{
\item{alleles}{one or more alleles, e.g. \code{DRB1_0101}.
See \link{get_netmhc2pan_alleles} for a full list}

\item{do_filter}{set to TRUE if the results of NetMHCIIpan must be filtered}

\item{fasta_filename}{the name of a FASTA file with protein sequences}

\item{folder_name}{the folder to install NetMHCIIpan, which is
\code{"/home/[usename]/.local/share"} by default}

\item{mhc_haplotype}{one MHC haplotype, e.g. \code{DRB1_0101}.
See \link{get_netmhc2pan_alleles} for a full list}

\item{netmhc2pan_archive_filename}{the NetMHC2pan archive
filename, for example \code{netMHCIIpan-3.2.Linux.tar.gz},
which is used in installation of NetMHC2pan.
Use \link{get_netmhc2pan_archive_filename} to get this filename.}

\item{netmhc2pan_bin_tarfile_path}{path of the
NetMHCIIpan binary tarball file}

\item{netmhc2pan_bin_url}{URL to download the
NetMHCIIpan binary tarball file from,
similar to, for example,
\code{https://www.cbs.dtu.dk/download/12345678-1234-1234-1234-123456789ABC}.
Use \link{get_netmhc2pan_bin_url} to get the default URL.}

\item{netmhc2pan_data_tarfile_path}{path of the
NetMHCIIpan data tarball file}

\item{netmhc2pan_data_url}{URL to download the
NetMHCIIpan data tarball file from,
similar to, for example,
\url{https://www.cbs.dtu.dk/services/NetMHCIIpan-3.2/data.Linux.tar.gz}.
Use \link{get_netmhc2pan_data_url} to get the default URL.}

\item{netmhc2pan_folder_name}{the folder (to be) used by NetMHCIIpan.
From this location, a subfolder for NetMHCIIpan is created.
Use \link{get_default_netmhc2pan_folder} to see the location
of the default NetMHCIIpan folder.
Use \link{get_default_netmhc2pan_subfolder} to see the location
of the default NetMHCIIpan subfolder.}

\item{netmhc2pan_subfolder}{the subfolder (to be) used by NetMHCIIpan.
Use \link{get_default_netmhc2pan_subfolder} to see the location
of the default NetMHCIIpan subfolder}

\item{netmhc2pan_version}{the NetMHCIIpan version,
for example \code{3.2}.
Use \link{get_default_netmhc2pan_version} to get the default
NetMHCIIpan version}

\item{os}{the operating system as obtained by \code{rappdirs::app_dir()$os}.
\code{netmhc2pan} supports Linux ('\code{unix}') only}

\item{peptide_length}{length of a peptide}

\item{peptides}{one or more peptide sequences}

\item{protein_sequence}{a protein sequence, for example \code{FAMILYVW}}

\item{temp_fasta_filename}{name for a temporary FASTA file,
which will be deleted automatically}

\item{temp_local_file}{path to the temporary file to store the URL to.
This file will be deleted afterwards.}

\item{temp_xls_filename}{name for a temporary \code{xls} file,
which will be deleted automatically}

\item{verbose}{set to TRUE for more output}

\item{xls_filename}{name of an \code{xls} file}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
