% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_netmhc2pan_bin_tarfile_path.R
\name{get_default_netmhc2pan_bin_tarfile_path}
\alias{get_default_netmhc2pan_bin_tarfile_path}
\title{Get the default path for the NetMHCIIpan binary tarball file}
\usage{
get_default_netmhc2pan_bin_tarfile_path(
  netmhc2pan_folder_name = get_default_netmhc2pan_folder(),
  netmhc2pan_archive_filename = get_netmhc2pan_archive_filename()
)
}
\arguments{
\item{netmhc2pan_folder_name}{the folder (to be) used by NetMHCIIpan.
From this location, a subfolder for NetMHCIIpan is created.
Use \link{get_default_netmhc2pan_folder} to see the location
of the default NetMHCIIpan folder.
Use \link{get_default_netmhc2pan_subfolder} to see the location
of the default NetMHCIIpan subfolder.}

\item{netmhc2pan_archive_filename}{the NetMHC2pan archive
filename, for example \code{netMHCIIpan-3.2.Linux.tar.gz},
which is used in installation of NetMHC2pan.
Use \link{get_netmhc2pan_archive_filename} to get this filename.}
}
\description{
Get the default path for the NetMHCIIpan binary tarball file.
This is the location where it will be saved to after downloading.
}
\examples{
get_default_netmhc2pan_bin_tarfile_path()
}
\author{
Richèl J.C. Bilderbeek
}
