% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_netmhc2pan_folder.R
\name{get_default_netmhc2pan_folder}
\alias{get_default_netmhc2pan_folder}
\title{Get the path to the folder where this package installs
NetMHCIIpan by default}
\usage{
get_default_netmhc2pan_folder()
}
\value{
the path to the folder where this package installs
  NetMHCIIpan by default
}
\description{
Get the path to the folder where this package installs
NetMHCIIpan by default
}
\examples{
get_default_netmhc2pan_folder()
}
\author{
Richèl J.C. Bilderbeek
}
