% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netmhc2pan_alleles.R
\name{get_netmhc2pan_alleles}
\alias{get_netmhc2pan_alleles}
\title{Get a list of the alleles supported by NetMHCIIpan}
\usage{
get_netmhc2pan_alleles(
  netmhc2pan_folder_name = get_default_netmhc2pan_folder()
)
}
\arguments{
\item{netmhc2pan_folder_name}{the folder (to be) used by NetMHCIIpan.
From this location, a subfolder for NetMHCIIpan is created.
Use \link{get_default_netmhc2pan_folder} to see the location
of the default NetMHCIIpan folder.
Use \link{get_default_netmhc2pan_subfolder} to see the location
of the default NetMHCIIpan subfolder.}
}
\value{
a character vector with the NetMHCIIpan alleles
}
\description{
Get a list of the alleles supported by NetMHCIIpan
}
\examples{
if (is_netmhc2pan_installed()) {
  get_netmhc2pan_alleles()[1:5]
}
}
\author{
Richèl J.C. Bilderbeek
}
