% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netmhc2pan_archive_filename.R
\name{get_netmhc2pan_archive_filename}
\alias{get_netmhc2pan_archive_filename}
\title{Get the filename of the NetMHC2pan archive file}
\usage{
get_netmhc2pan_archive_filename(
  netmhc2pan_version = get_default_netmhc2pan_version()
)
}
\arguments{
\item{netmhc2pan_version}{the NetMHCIIpan version,
for example \code{3.2}.
Use \link{get_default_netmhc2pan_version} to get the default
NetMHCIIpan version}
}
\value{
the filename of the NetMHC2pan archive file
}
\description{
Get the filename of the NetMHC2pan archive file
}
\examples{
get_netmhc2pan_archive_filename()
}
