% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netmhc2pan_bin_url.R
\name{get_netmhc2pan_bin_url}
\alias{get_netmhc2pan_bin_url}
\title{Get the NetMHCIIpan binary download URL.}
\usage{
get_netmhc2pan_bin_url()
}
\value{
a download URL
}
\description{
Get the URL for the NetMHCIIpan binary, as is emailed.
These expire after 4 hours.
}
\seealso{
Use \link{is_url_valid} to determine if the download URL is still valid.
Use \link{check_netmhc2pan_bin_url} to get a helpful error message if this
URL is invalid.
}
\author{
Richèl J.C. Bilderbeek
}
