% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netmhc2pan_data_url.R
\name{get_netmhc2pan_data_url}
\alias{get_netmhc2pan_data_url}
\title{Get the NetMHCIIpan data tarball URL.}
\usage{
get_netmhc2pan_data_url()
}
\value{
a download URL
}
\description{
Get the NetMHCIIpan data tarball URL.
}
\seealso{
Use \link{is_url_valid} to determine if the download URL is still valid.
Use \link{check_netmhc2pan_data_url} to get a helpful error message if this
URL is invalid.
}
\author{
Richèl J.C. Bilderbeek
}
