% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhc2pan_bin_from_file.R
\name{install_netmhc2pan_bin_from_file}
\alias{install_netmhc2pan_bin_from_file}
\title{Install the NetMHCIIpan binary to a local folder}
\usage{
install_netmhc2pan_bin_from_file(
  netmhc2pan_bin_tarfile_path,
  verbose = FALSE,
  netmhc2pan_archive_filename = get_netmhc2pan_archive_filename(),
  netmhc2pan_folder_name = get_default_netmhc2pan_folder()
)
}
\arguments{
\item{netmhc2pan_bin_tarfile_path}{path of the
NetMHCIIpan binary tarball file}

\item{verbose}{set to TRUE for more output}

\item{netmhc2pan_archive_filename}{the NetMHC2pan archive
filename, for example \code{netMHCIIpan-3.2.Linux.tar.gz},
which is used in installation of NetMHC2pan.
Use \link{get_netmhc2pan_archive_filename} to get this filename.}

\item{netmhc2pan_folder_name}{the folder (to be) used by NetMHCIIpan.
From this location, a subfolder for NetMHCIIpan is created.
Use \link{get_default_netmhc2pan_folder} to see the location
of the default NetMHCIIpan folder.
Use \link{get_default_netmhc2pan_subfolder} to see the location
of the default NetMHCIIpan subfolder.}
}
\value{
Nothing
}
\description{
Install the NetMHCIIpan binary to a local folder
}
\author{
Richèl J.C. Bilderbeek
}
