% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhc2pan_from_files.R
\name{install_netMHCpan_from_files}
\alias{install_netMHCpan_from_files}
\title{Install NetMHCIIpan from files.}
\usage{
install_netMHCpan_from_files(
  netMHCpan_bin_tarfile_path,
  netMHCpan_data_tarfile_path,
  verbose = FALSE,
  netMHCpan_archive_filename = get_netMHCpan_archive_filename(),
  netMHCpan_folder_name = get_default_netMHCpan_folder(),
  temp_local_file = tempfile(pattern = "netMHCpan_install_netMHCpan_")
)
}
\arguments{
\item{netMHCpan_bin_tarfile_path}{path of the
  NetMHCpan binary tarball file
##This needs CHANGING}

\item{netMHCpan_data_tarfile_path}{path of the
NetMHCpan data tarball file}

\item{verbose}{set to TRUE for more output}

\item{netMHCpan_archive_filename}{the NetMHCpan archive
filename, for example \code{netMHCpan-4.1.Linux.tar.gz},
which is used in installation of NetMHCpan.
Use \link{get_netMHCpan_archive_filename} to get this filename.}

\item{netMHCpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netMHCpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netMHCpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{temp_local_file}{path to the temporary file to store the URL to.
This file will be deleted afterwards.}
}
\value{
Nothing
}
\description{
Install NetMHCIIpan to a local folder from (already downloaded)
binary and date tarball files.
Use \link{install_netmhc2pan} to install
NetMHCIIpan by downloading these files
}
\details{
These are three steps:

\itemize{
  \item Install the NetMHCIIpan binary,
    using \link{install_netmhc2pan_bin_from_file}
  \item Install the NetMHCIIpan data,
    using \link{install_netmhc2pan_data_from_file}
  \item Set up NetMHCIIpan, using \link{set_up_netmhc2pan}
}
}
\author{
Richèl J.C. Bilderbeek
}
