% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_netmhc2pan_installed.R
\name{is_netmhc2pan_data_installed}
\alias{is_netmhc2pan_data_installed}
\title{Measure if NetMHCIIpan data folder is installed locally}
\usage{
is_netmhc2pan_data_installed(
  netmhc2pan_folder_name = get_default_netmhc2pan_folder()
)
}
\arguments{
\item{netmhc2pan_folder_name}{the folder (to be) used by NetMHCIIpan.
From this location, a subfolder for NetMHCIIpan is created.
Use \link{get_default_netmhc2pan_folder} to see the location
of the default NetMHCIIpan folder.
Use \link{get_default_netmhc2pan_subfolder} to see the location
of the default NetMHCIIpan subfolder.}
}
\value{
TRUE is NetMHCIIpan data folder is installed locally,
  FALSE otherwise
}
\description{
Measure if NetMHCIIpan data folder is installed locally
}
\author{
Richèl J.C. Bilderbeek
}
