% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_netmhc2pan_installed.R
\name{is_netmhc2pan_installed}
\alias{is_netmhc2pan_installed}
\title{Measure if NetMHCIIpan is installed locally}
\usage{
is_netmhc2pan_installed(
  netmhc2pan_folder_name = get_default_netmhc2pan_folder(),
  verbose = FALSE
)
}
\arguments{
\item{netmhc2pan_folder_name}{the folder (to be) used by NetMHCIIpan.
From this location, a subfolder for NetMHCIIpan is created.
Use \link{get_default_netmhc2pan_folder} to see the location
of the default NetMHCIIpan folder.
Use \link{get_default_netmhc2pan_subfolder} to see the location
of the default NetMHCIIpan subfolder.}

\item{verbose}{set to TRUE for more output}
}
\value{
TRUE is NetMHCIIpan is installed locally,
  FALSE otherwise
}
\description{
Measure if NetMHCIIpan is installed locally
}
\author{
Richèl J.C. Bilderbeek
}
