% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_netmhc2pan_installed.R
\name{is_netMHCpan_bin_installed}
\alias{is_netMHCpan_bin_installed}
\title{Measure if NetMHCIIpan binary is installed locally}
\usage{
is_netMHCpan_bin_installed(
  netMHCpan_folder_name = get_default_netMHCpan_folder(),
  verbose = FALSE
)
}
\arguments{
\item{netMHCpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netMHCpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netMHCpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{verbose}{set to TRUE for more output}
}
\value{
TRUE is NetMHCIIpan binary is installed locally,
  FALSE otherwise
}
\description{
Measure if NetMHCIIpan binary is installed locally
}
\author{
Richèl J.C. Bilderbeek
}
