% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_netmhc2pan_installed.R
\name{is_tcsh_installed}
\alias{is_tcsh_installed}
\title{See if tcsh is installed}
\usage{
is_tcsh_installed()
}
\value{
TRUE is tcsh is installed,
  FALSE otherwise
}
\description{
See if tcsh is installed
}
\note{
To install \code{tcsh} under Linux, do \code{sudo apt install tcsh}
}
\author{
Richèl J.C. Bilderbeek
}
