% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_url_valid.R
\name{is_url_valid}
\alias{is_url_valid}
\title{Is the download URL valid?}
\usage{
is_url_valid(
  netmhc2pan_bin_url = get_netmhc2pan_bin_url(),
  verbose = FALSE,
  netmhc2pan_archive_filename = get_netmhc2pan_archive_filename(),
  temp_local_file = tempfile(pattern = "netmhc2pan_is_url_valid_")
)
}
\arguments{
\item{netmhc2pan_bin_url}{URL to download the
NetMHCIIpan binary tarball file from,
similar to, for example,
\code{https://www.cbs.dtu.dk/download/12345678-1234-1234-1234-123456789ABC}.
Use \link{get_netmhc2pan_bin_url} to get the default URL.}

\item{verbose}{set to TRUE for more output}

\item{netmhc2pan_archive_filename}{the NetMHC2pan archive
filename, for example \code{netMHCIIpan-3.2.Linux.tar.gz},
which is used in installation of NetMHC2pan.
Use \link{get_netmhc2pan_archive_filename} to get this filename.}

\item{temp_local_file}{path to the temporary file to store the URL to.
This file will be deleted afterwards.}
}
\description{
The download link expires after 4 hours.
}
