% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netmhc2pan_self_test.R
\name{netmhc2pan_self_test}
\alias{netmhc2pan_self_test}
\title{Test NetMHCIIpan}
\usage{
netmhc2pan_self_test(
  netmhc2pan_folder_name = get_default_netmhc2pan_folder(),
  verbose = FALSE
)
}
\arguments{
\item{netmhc2pan_folder_name}{the folder (to be) used by NetMHCIIpan.
From this location, a subfolder for NetMHCIIpan is created.
Use \link{get_default_netmhc2pan_folder} to see the location
of the default NetMHCIIpan folder.
Use \link{get_default_netmhc2pan_subfolder} to see the location
of the default NetMHCIIpan subfolder.}

\item{verbose}{set to TRUE for more output}
}
\description{
Test NetMHCIIpan
}
\author{
Richèl J.C. Bilderbeek
}
