% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_netmhc2pan.R
\name{set_up_netmhc2pan}
\alias{set_up_netmhc2pan}
\title{Install the NetMHCIIpan binary to a local folder}
\usage{
set_up_netmhc2pan(
  netmhc2pan_folder_name = get_default_netmhc2pan_folder(),
  verbose = FALSE
)
}
\arguments{
\item{netmhc2pan_folder_name}{the folder (to be) used by NetMHCIIpan.
From this location, a subfolder for NetMHCIIpan is created.
Use \link{get_default_netmhc2pan_folder} to see the location
of the default NetMHCIIpan folder.
Use \link{get_default_netmhc2pan_subfolder} to see the location
of the default NetMHCIIpan subfolder.}

\item{verbose}{set to TRUE for more output}
}
\value{
Nothing
}
\description{
Install the NetMHCIIpan binary to a local folder
}
\examples{
\donttest{
  set_up_netmhc2pan()
}
}
\author{
Richèl J.C. Bilderbeek
}
